pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "macro-trading"
version = "0.1.0"
description = "Macro Trading Data Infrastructure - volatility signals and market sentiment analysis"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "sqlalchemy>=2.0.36,<2.1",
    "asyncpg>=0.30.0,<0.31",
    "psycopg2-binary>=2.9.10,<2.10",
    "alembic>=1.14.0,<1.15",
    "pydantic-settings>=2.7.0,<2.8",
    "pydantic>=2.10.0,<2.11",
    "redis[hiredis]>=5.2.0,<5.3",
    "python-dotenv>=1.0.1,<1.1",
    "structlog>=24.4.0,<25.0",
    "httpx>=0.27.0,<0.28",
    "tenacity>=8.2.0,<8.3",
    "yfinance>=0.2.36,<0.3",
    "bizdays>=1.0.0,<1.1",
    "exchange_calendars>=4.5.0,<4.6",
    "python-dateutil>=2.8.0,<2.9",
    "pandas>=2.1,<2.3",
    "numpy>=1.26,<2.0",
    "scipy>=1.12,<1.15",
    "statsmodels>=0.14,<0.15",
    "scikit-learn>=1.4,<1.6",
    "matplotlib>=3.8,<3.10",
    "slowapi>=0.1.9,<0.2",
    "PyJWT>=2.8.0,<2.10",
    "fastapi>=0.115.0,<0.116",
    "uvicorn>=0.32.0,<0.33",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
    "respx>=0.21.0",
    "pytest-cov>=4.1.0",
    "bandit>=1.7.0",
    "pip-audit>=2.7.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
